backend/pyproject.toml
[project]
name = "nats-jetstream-manager"
version = "1.0.0"
description = "FastAPI backend for managing NATS JetStream clusters"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]==0.115.0",
    "uvicorn[standard]==0.32.0",
    "nats-py==2.7.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "black==24.10.0",
    "ruff==0.7.4",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
